Move logic to update searchedAddress and coordinates from searchParams to AddressMapper component. Remove useEffect and memoized functions in SearchBar component. Remove unused props.
