+++ app/components/home-header.tsx
@@ -20,7 +20,6 @@ export type HazardData = {
 };
 
 interface HomeHeaderProps {
-  coordinates: number[];
   searchedAddress: string;
   onSearchChange: (coords: number[]) => void;
   onAddressSearch: (address: string) => void;
@@ -31,7 +30,6 @@ interface HomeHeaderProps {
 const SEARCHBAR_PORTAL_ID = "searchbar-portal";
 
 const HomeHeader = ({
-  coordinates,
   searchedAddress,
   onSearchChange,
   onAddressSearch,
@@ -92,7 +90,6 @@ const HomeHeader = ({
       {/* NOTE: This Suspense boundary is being used around a component that utilizes `useSearchParams()` to prevent entire page from deopting into client-side rendering (CSR) bailout as per https://nextjs.org/docs/messages/missing-suspense-with-csr-bailout */}
       <Suspense fallback={<SearchBarSkeleton />}>
         <SearchBar
-          coordinates={coordinates}
           onSearchChange={onSearchChange}
           onAddressSearch={onAddressSearch}
           onCoordDataRetrieve={onCoordDataRetrieve}
